Backend/Python/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.1.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.7
structlog==23.1.0
boto3
camel_converter